# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Monitor API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [procurement]
      summary: List purchase orders (scoped to user role/sector)
      parameters:
        - name: sectorId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, submitted, approved, rejected]
        - name: cycleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrderWithDetails"

    post:
      operationId: createPurchaseOrder
      tags: [procurement]
      summary: Create a new draft purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderBody"
      responses:
        "201":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithDetails"
  /purchase-orders/{orderId}:
    get:
      operationId: getPurchaseOrder
      tags: [procurement]
      summary: Get a purchase order with its line items
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithDetails"
  /purchase-orders/{orderId}/submit:
    post:
      operationId: submitPurchaseOrder
      tags: [procurement]
      summary: Submit a draft order for approval
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithDetails"
  /purchase-orders/{orderId}/review:
    post:
      operationId: reviewPurchaseOrder
      tags: [procurement]
      summary: Approve or reject a submitted purchase order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewOrderBody"
      responses:
        "200":
          description: Order reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithDetails"
  /purchase-orders/{orderId}/items:
    post:
      operationId: addOrderItem
      tags: [procurement]
      summary: Add a line item to a draft purchase order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderItemBody"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithDetails"
  /purchase-orders/{orderId}/items/{itemId}:
    delete:
      operationId: removeOrderItem
      tags: [procurement]
      summary: Remove a line item from a draft order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithDetails"

  # REPORTS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, ceo, ministry_head, department_head, viewer]
        sectorId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    Sector:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        parentId:
          type: integer
          nullable: true
        depth:
          type: integer
        responsibleUserId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - depth
        - isActive
        - sortOrder
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        unitPrice:
          type: number
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - unit
        - unitPrice
        - isActive
        - sortOrder
        - createdAt
...
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: number
        unitPriceSnapshot:
          type: number
        lineTotal:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - productId
        - quantity
        - unitPriceSnapshot
        - lineTotal
        - createdAt
    PurchaseOrderWithDetails:
      type: object
      properties:
        id:
          type: integer
        sectorId:
          type: integer
        sector:
          $ref: "#/components/schemas/Sector"
        budgetCycleId:
          type: integer
        createdBy:
          type: integer
        createdByUser:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [draft, submitted, approved, rejected]
        notes:
          type: string
          nullable: true
        totalAmount:
          type: number
        submittedAt:
          type: string
          format: date-time
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewedBy:
          type: integer
          nullable: true
        reviewedByUser:
          $ref: "#/components/schemas/User"
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sectorId
        - budgetCycleId
        - createdBy
        - status
        - totalAmount
        - items
        - createdAt
    CreatePurchaseOrderBody:
      type: object
      properties:
        sectorId:
          type: integer
        budgetCycleId:
          type: integer
        notes:
          type: string
      required:
        - sectorId
        - budgetCycleId
    AddOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        notes:
          type: string
      required:
        - productId
        - quantity
    ReviewOrderBody:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
        rejectionReason:
          type: string
      required:
        - action
...